Pascal case from snake_case for proper field names
