+++ pkg/codex/docs/openapi.md
@@ -1,0 +1,9 @@
+# OpenApi
+
+## References
+- https://www.openapis.org/
+
+## Resources
+- https://github.com/getkin/kin-openapi/openapi3 -  Go Library for OpenApi 3
+- https://github.com/deepmap/oapi-codegen - Repo that uses the kin-openapi 
+- https://github.com/teamdigitale/api-openapi-samples
