+++ semver
@@ -1,3 +1,7 @@
+0.0.5
+
+- Pascal names from snake_case
+
 0.0.4
 
 - update the go template docs
